requirements.txt
fastapi
uvicorn
python-dotenv
pandas
requests
langchain
openai
langchain-openai
langgraph
langchain-community
langchain-google-genai
langchain-chroma
langchain-groq
chromadb